requirements.txt
asgiref==3.11.0
channels==4.1.0
channels-redis==4.2.0
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
msgpack==1.1.2
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
redis==7.1.0
sqlparse==0.5.5
tzdata==2025.3
